fix build, sync with our working copy.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@5279 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/libteletone/src/libteletone.h
@@ -78,7 +78,7 @@ extern "C" {
 #define TELETONE_MAX_TONES 6
 #define TELETONE_TONE_RANGE 127
 
-typedef float teletone_process_t;
+typedef double teletone_process_t;
 
 /*! \file libteletone.h
     \brief Top level include file
